refactor(ui): update shadcn primitives import cn from lib/cn
