+++ ui/src/components/ui/async-select.tsx
@@ -6,7 +6,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select"
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/cn"
 
 function SelectSkeleton({ className }: { className?: string }) {
   return (
