+++ ui/src/components/ui/badge.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 import { cva, type VariantProps } from "class-variance-authority"
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/cn"
 
 const badgeVariants = cva(
   [
