+++ ui/src/components/ui/label.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 import * as LabelPrimitive from "@radix-ui/react-label"
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/cn"
 
 const Label = React.forwardRef<
   React.ElementRef<typeof LabelPrimitive.Root>,
