+++ ui/src/components/ui/select.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import * as SelectPrimitive from "@radix-ui/react-select"
 import { Check, ChevronDown } from "lucide-react"
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/cn"
 
 const Select = SelectPrimitive.Root
 const SelectGroup = SelectPrimitive.Group
